refactor(ChatPane): extract tool-categorization into pure toolCallCategorization.ts
Moves the pure categorization/summarization cluster (categorizeItem/ToolName/
CommandExecution/VerbPrefix/PersistedSummary, summarizeToolCalls/SameFileEditGroup,
GroupCategory/CategoryMeta/CATEGORY_META/SUMMARY_CATEGORY_LABELS, isToolGroupItem/
isToolLikeItem, normalizeEditGroupPath/readEditGroupPath) plus the shared payload
readers (getToolLikePayload, readCommandPayloadCommand, isEditLikeToolPayload) out
of ToolCallGroup.tsx into a React-free toolCallCategorization.ts, with a new
focused unit test (12 cases). ToolCallGroup.tsx drops 965->660 and imports the
pure module one-directionally (no cycle). Behavior-preserving; the row builders +
their helpers stay in the .tsx. ToolCallGroup.test.tsx (23) green.
